Refactor GPS capture
