+++ CMakeLists.txt
@@ -2,10 +2,11 @@
 
 cmake_minimum_required(VERSION 3.13)
 
+
 set(CMAKE_C_STANDARD 11)
 set(CMAKE_CXX_STANDARD 17)
 
-# initalize pico_sdk from installed location
+# Initialise pico_sdk from installed location
 # (note this can come from environment, CMake cache etc)
 set(PICO_SDK_PATH "/home/pi/pico/pico-sdk")
 
@@ -17,7 +18,7 @@ project(rockblock_gps C CXX ASM)
 # Initialise the Raspberry Pi Pico SDK
 pico_sdk_init()
 
-# Add executable. Default name is the project name, version 0.3
+# Add executable. Default name is the project name, version 0.1
 
 add_subdirectory(serial)
 add_subdirectory(rockblock)
