+++ rockblock_gps/rockblock_gps.hpp
@@ -13,26 +13,10 @@
 
 #include "pico/stdlib.h"
 
-#define UART1_ID uart1
-#define UART1_TX_PIN 4
-#define UART1_RX_PIN 5
+// const char* construct_message(char *message_buf, char* message, int message_count);
 
-#define RB_BAUD_RATE 19200
-#define GPS_BAUD_RATE 9600
+// void led_init();
 
-#define OUTPUT_INTERVAL 5000 // ms
-
-#define BUTTON_PIN 15
-#define LED_AMBER 14
-#define LED_RED 16
-#define LED_GREEN 18
-#define LED_PIN PICO_DEFAULT_LED_PIN
-
-
-const char* construct_message(char *message_buf, char* message, int message_count);
-
-void led_init();
-
-void switch_led(uint state);
+// void switch_led(uint state);
 
 #endif
